+++ packages/renderer/src/view/index.ts
@@ -1,0 +1,4 @@
+export type { RenderEntityView } from './RenderEntityView'
+export type { RenderWorldView } from './RenderWorldView'
+export type { PixiEntityRenderer, PixiEntityRendererOptions } from './PixiEntityRenderer'
+export { DefaultPixiEntityRenderer } from './PixiEntityRenderer'